% Generated by roxygen2: do not edit by hand
% Please edit documentation in create_mezen.Rmd.
\name{analyse_records}
\alias{analyse_records}
\title{analyses the records within the given list of community records and extracts the metainformation from it}
\usage{
analyse_records(gmci_records_id)
}
\arguments{
\item{gmci_records_id}{Name of a person}
}
\description{
analyses the records within the given list of community records and extracts the metainformation from it
}