% Generated by roxygen2: do not edit by hand
% Please edit documentation in create_mezen.Rmd.
\name{download_data}
\alias{download_data}
\title{Question to the LLM}
\usage{
download_data(question_text, json_file = "example.json")
}
\arguments{
\item{question_text}{parameter that sets the input text}
}
\description{
Question to the LLM
}