% Generated by roxygen2: do not edit by hand
% Please edit documentation in create_mezen.Rmd.
\name{load_community_records}
\alias{load_community_records}
\title{Download the necessary files}
\usage{
load_community_records(name, exclamation = TRUE)
}
\arguments{
\item{name}{parameter that sets the name of the repository}
}
\description{
Download the necessary files
}