% Generated by roxygen2: do not edit by hand
% Please edit documentation in create_mezen.Rmd.
\name{question}
\alias{question}
\title{Asks LLM to retrieve the information}
\usage{
question(question_text, json_file = "example.json")
}
\arguments{
\item{question_text}{parameter that define the questions that is targeted towards the LLM}
}
\description{
Asks LLM to retrieve the information
}